+++ node/clients/index.ts
@@ -1,16 +1,16 @@
 import { IOClients } from '@vtex/api'
 
 import { Checkout } from './checkout'
+import { SearchGraphQL } from './searchGraphQL'
 import { Shipping } from './shipping'
-import { StoreGraphQL } from './storeGraphQL'
 
 export class Clients extends IOClients {
   public get checkout() {
     return this.getOrSet('checkout', Checkout)
   }
 
-  public get storeGraphQL() {
-    return this.getOrSet('storeGraphQL', StoreGraphQL)
+  public get searchGraphQL() {
+    return this.getOrSet('searchGraphQL', SearchGraphQL)
   }
 
   public get shipping() {